+++ src/ts/ConvivaAnalytics.ts
@@ -1,5 +1,5 @@
 import * as Conviva from '@convivainc/conviva-js-coresdk';
-import type {
+import {
   AdBreakEvent,
   AdEvent,
   AudioChangedEvent,
@@ -19,12 +19,19 @@ import { ConvivaAnalyticsConfiguration, ConvivaAnalyticsTracker, EventAttributes
 import { ConvivaAnalyticsSsai } from './ConvivaAnalyticsSsai';
 import { PlayerEventWrapper } from './helper/PlayerEventWrapper';
 import { AdHelper } from './helper/AdHelper';
+import { Html5Logging } from './Html5Logging';
 
 export class ConvivaAnalytics {
-  private readonly events: typeof PlayerEvent;
-  private readonly handlers: PlayerEventWrapper;
+  private handlers?: PlayerEventWrapper;
   private readonly convivaAnalyticsTracker: ConvivaAnalyticsTracker;
-  private readonly player: PlayerAPI;
+  private _player?: PlayerAPI;
+
+  private get player(): PlayerAPI {
+    if (!this._player) {
+      throw new Error('Player is not initialized, either pass it to the constructor or attach it via `attachPlayer` before using the integration.');
+    }
+    return this._player;
+  }
 
   private readonly debugLoggingEnabled: boolean;
 
@@ -36,17 +43,18 @@ export class ConvivaAnalytics {
 
   private convivaSsaiAnalytics: ConvivaAnalyticsSsai;
 
+  private readonly logger: Conviva.LoggingInterface = new Html5Logging();
+
   public readonly ssai: Omit<ConvivaAnalyticsSsai, 'reset'>;
 
-  constructor(player: PlayerAPI, customerKey: string, config: ConvivaAnalyticsConfiguration = {}) {
-    this.convivaAnalyticsTracker = new ConvivaAnalyticsTracker(player, customerKey, config);
+  constructor(player: PlayerAPI | undefined, customerKey: string, config: ConvivaAnalyticsConfiguration = {}) {
+    this.convivaAnalyticsTracker = new ConvivaAnalyticsTracker(customerKey, config);
     this.debugLoggingEnabled = config.debugLoggingEnabled || false;
-    this.player = player;
-    // TODO: Use alternative to deprecated player.exports
-    this.events = player.exports.PlayerEvent;
-    this.handlers = new PlayerEventWrapper(player);
+    this._player = player;
 
-    this.registerPlayerEvents();
+    if (player) {
+      this.attachPlayer(player);
+    }
 
     this.convivaSsaiAnalytics = new ConvivaAnalyticsSsai(this.convivaAnalyticsTracker);
 
@@ -61,6 +69,46 @@ export class ConvivaAnalytics {
       reportAdSkipped: this.convivaSsaiAnalytics.reportAdSkipped.bind(this.convivaSsaiAnalytics),
       reportAdBreakFinished: this.convivaSsaiAnalytics.reportAdBreakFinished.bind(this.convivaSsaiAnalytics),
     };
+  }
+
+  /**
+   * Attaches the player instance to the integration. This can be used for late player attaching e.g.
+   * to measure VST (Video start time) more precisely.
+   *
+   * Has no effect if there is already a `Player` instance set. Use the `new ConvivaAnalytics(...)` without `player`
+   * if you plan to attach a `Player` instance later in the life-cycle.
+   *
+   * Example:
+   * ```
+   * const convivaAnalytics = new ConvivaAnalytics(undefined, '<CONVIVA_CUSTOMER_KEY>')
+   *
+   * // Asset name is required to be set when initializing the session before `player.load()`.
+   * convivaAnalytics.updateContentMetadata({ assetName: 'My video' });
+   * convivaAnalytics.initializeSession();
+   *
+   * // ... Additional setup steps
+   *
+   * convivaAnalytics.attachPlayer(player);
+   * player.load({ ... });
+   * ```
+   *
+   * @param player The player instance to attach to the integration.
+   */
+  public attachPlayer(player: PlayerAPI): void {
+    const {canAttach, reason} = this.convivaAnalyticsTracker.canAttachPlayer(player);
+
+    if (!canAttach) {
+      this.logger.consoleLog(
+        `[ ConvivaAnalyticsTracker ] cannot attach player: ${reason}`,
+        Conviva.SystemSettings.LogLevel.WARNING,
+      );
+      return;
+    }
+
+    this.convivaAnalyticsTracker.attachPlayer(player);
+    this._player = player;
+    this.handlers = new PlayerEventWrapper(player);
+    this.registerPlayerEvents();
   }
 
   /**
@@ -87,13 +135,9 @@ export class ConvivaAnalytics {
    * no longer ensure that the session is managed at the correct time.
    */
   public endSession(): void {
-    this.reset();
-    this.convivaAnalyticsTracker.endSession();
-  }
-
-  private reset(): void {
-    this.lastAdBreakEvent = null;
+    this.debugLog('[ ConvivaAnalytics ] external ending session');
     this.convivaSsaiAnalytics.reset();
+    this.convivaAnalyticsTracker.endSession();
   }
 
   /**
@@ -161,14 +205,26 @@ export class ConvivaAnalytics {
   }
 
   public release(): void {
-    this.destroy();
-    this.convivaAnalyticsTracker.release();
+    this.releaseInternal();
   }
 
-  private destroy(event?: PlayerEventBase): void {
-    this.reset();
-    this.unregisterPlayerEvents();
-    this.convivaAnalyticsTracker.release(event);
+  private releaseInternal(event?: PlayerEventBase): void {
+    const isPlayerDestroyed = event?.type === PlayerEvent.Destroy;
+
+    if (!isPlayerDestroyed) {
+      this.unregisterPlayerEvents();
+    }
+
+    this._player = null;
+    this.handlers = null;
+
+    this.debugLog('[ ConvivaAnalytics ] releasing', {
+      event,
+      isPlayerDestroyed,
+    });
+    this.convivaAnalyticsTracker.release(isPlayerDestroyed);
+    this.convivaSsaiAnalytics.reset();
+    this.lastAdBreakEvent = null;
   }
 
   private debugLog(message?: any, ...optionalParams: any[]): void {
@@ -297,42 +353,42 @@ export class ConvivaAnalytics {
 
   private onDestroy = (event: any) => {
     this.debugLog('[ ConvivaAnalytics ] [ Player Event ] destroy', event);
-    this.destroy(event);
+    this.releaseInternal(event);
   };
 
   private registerPlayerEvents(): void {
-    this.handlers.add(this.events.Play, this.onPlay);
-    this.handlers.add(this.events.Playing, this.onPlaying);
-    this.handlers.add(this.events.Paused, this.onPlaybackStateChanged);
-    this.handlers.add(this.events.StallStarted, this.onPlaybackStateChanged);
-    this.handlers.add(this.events.StallEnded, this.onPlaybackStateChanged);
-    this.handlers.add(this.events.PlaybackFinished, this.onPlaybackFinished);
-    this.handlers.add(this.events.VideoPlaybackQualityChanged, this.onVideoQualityChanged);
-    this.handlers.add(this.events.AudioPlaybackQualityChanged, this.onCustomEvent);
-    this.handlers.add(this.events.Muted, this.onCustomEvent);
-    this.handlers.add(this.events.Unmuted, this.onCustomEvent);
-    this.handlers.add(this.events.ViewModeChanged, this.onCustomEvent);
-    this.handlers.add(this.events.AdStarted, this.onAdStarted);
-    this.handlers.add(this.events.AdFinished, this.onAdFinished);
-    this.handlers.add(this.events.AdBreakStarted, this.onAdBreakStarted);
-    this.handlers.add(this.events.AdBreakFinished, this.onAdBreakFinished);
-    this.handlers.add(this.events.AdSkipped, this.onAdSkipped);
-    this.handlers.add(this.events.AdError, this.onAdError);
-    this.handlers.add(this.events.Error, this.onError);
-    this.handlers.add(this.events.Destroy, this.onDestroy);
-    this.handlers.add(this.events.Seek, this.onSeek);
-    this.handlers.add(this.events.Seeked, this.onSeeked);
-    this.handlers.add(this.events.TimeShift, this.onTimeShift);
-    this.handlers.add(this.events.TimeShifted, this.onTimeShifted);
-    this.handlers.add(this.events.AudioChanged, this.onAudioChanged);
-    this.handlers.add(this.events.SubtitleEnabled, this.onSubtitleEnabled);
-    this.handlers.add(this.events.SubtitleDisabled, this.onSubtitleDisabled);
+    this.handlers.add(PlayerEvent.Play, this.onPlay);
+    this.handlers.add(PlayerEvent.Playing, this.onPlaying);
+    this.handlers.add(PlayerEvent.Paused, this.onPlaybackStateChanged);
+    this.handlers.add(PlayerEvent.StallStarted, this.onPlaybackStateChanged);
+    this.handlers.add(PlayerEvent.StallEnded, this.onPlaybackStateChanged);
+    this.handlers.add(PlayerEvent.PlaybackFinished, this.onPlaybackFinished);
+    this.handlers.add(PlayerEvent.VideoPlaybackQualityChanged, this.onVideoQualityChanged);
+    this.handlers.add(PlayerEvent.AudioPlaybackQualityChanged, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.Muted, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.Unmuted, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.ViewModeChanged, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.AdStarted, this.onAdStarted);
+    this.handlers.add(PlayerEvent.AdFinished, this.onAdFinished);
+    this.handlers.add(PlayerEvent.AdBreakStarted, this.onAdBreakStarted);
+    this.handlers.add(PlayerEvent.AdBreakFinished, this.onAdBreakFinished);
+    this.handlers.add(PlayerEvent.AdSkipped, this.onAdSkipped);
+    this.handlers.add(PlayerEvent.AdError, this.onAdError);
+    this.handlers.add(PlayerEvent.Error, this.onError);
+    this.handlers.add(PlayerEvent.Destroy, this.onDestroy);
+    this.handlers.add(PlayerEvent.Seek, this.onSeek);
+    this.handlers.add(PlayerEvent.Seeked, this.onSeeked);
+    this.handlers.add(PlayerEvent.TimeShift, this.onTimeShift);
+    this.handlers.add(PlayerEvent.TimeShifted, this.onTimeShifted);
+    this.handlers.add(PlayerEvent.AudioChanged, this.onAudioChanged);
+    this.handlers.add(PlayerEvent.SubtitleEnabled, this.onSubtitleEnabled);
+    this.handlers.add(PlayerEvent.SubtitleDisabled, this.onSubtitleDisabled);
 
-    this.handlers.add(this.events.CastStarted, this.onCustomEvent);
-    this.handlers.add(this.events.CastStopped, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.CastStarted, this.onCustomEvent);
+    this.handlers.add(PlayerEvent.CastStopped, this.onCustomEvent);
   }
 
   private unregisterPlayerEvents(): void {
-    this.handlers.clear();
+    this.handlers?.clear();
   }
 }